requirements.txt
requests==2.29.0
ConfigArgParse==1.5.3
python-telegram-bot==13.15